refac #20: 타입 정리
