+++ components/social/feed.tsx
@@ -1,9 +1,9 @@
 import { getElapsedTime } from "@/lib/client/getElapsedTime";
 import useMutation from "@/lib/client/useMutation";
-import { IMemoWithReactions } from "@/pages/book/[bookId]";
 import { useSession } from "next-auth/react";
 import { useRouter } from "next/router";
 import Comment from "@/components/book/detail/memo/comment";
+import { IMemoWithReactions } from "@/types";
 
 interface IProps {
   memo: IMemoWithReactions;
